refactor(components): group by feature into subdirectories
16 files moved into 4 new subdirectories. 16 importers updated. No public
surface change beyond import paths.

New subdirs:
- block-sale/  (9 files: form, queue, stand-form, stand-panel, maps + tests)
- market/      (3 files: detail, image-gallery + test)
- booking/     (2 files: tables-card + test)
- route-view/  (1 file: route-view-map.tsx — renamed from root-level
                route-map.tsx to disambiguate from route-builder/route-map.tsx)

Existing dirs unchanged:
- route-builder/ — parent component moves into the dir
  (route-builder.tsx → route-builder/route-builder.tsx); the dynamic
  import path becomes @/components/route-builder/route-builder
- market-form/ — already correctly grouped

Files staying at root: address-card, address-picker, auto-imported-notice,
back-link, claim-market-button, cookie-* , fyndstigen-logo, fyndstigen-map,
image-upload-list, map-view, nav, opening-hours-*, organizer-card,
stripe-connect-button, track-market-view, trail-background,
web-vitals-reporter.

Closes #88
