+++ web/src/app/kvartersloppis/[slug]/admin/page.tsx
@@ -3,7 +3,7 @@
 import { useParams } from 'next/navigation'
 import { useAuth } from '@/lib/auth/auth-context'
 import { useBlockSale } from '@/hooks/use-block-sale'
-import { BlockSaleQueue } from '@/components/block-sale-queue'
+import { BlockSaleQueue } from '@/components/block-sale/queue'
 
 export default function AdminPage() {
   const params = useParams<{ slug: string }>()
