+++ web/src/app/loppis/[slug]/page.tsx
@@ -2,7 +2,7 @@ import { notFound } from 'next/navigation'
 import { permanentRedirect } from 'next/navigation'
 import { createSupabaseServerClient } from '@/lib/supabase/server'
 import { createSupabaseServerData } from '@fyndstigen/shared'
-import { MarketDetail } from '@/components/market-detail'
+import { MarketDetail } from '@/components/market/detail'
 import { TrackMarketView } from '@/components/track-market-view'
 
 type Props = { params: Promise<{ slug: string }> }
