+++ web/src/app/skapa/kvartersloppis/page.tsx
@@ -4,7 +4,7 @@ import { useEffect } from 'react'
 import { useRouter } from 'next/navigation'
 import { usePostHog } from 'posthog-js/react'
 import { useAuth } from '@/lib/auth/auth-context'
-import { BlockSaleForm } from '@/components/block-sale-form'
+import { BlockSaleForm } from '@/components/block-sale/form'
 import { useBlockSaleCreate } from '@/hooks/use-block-sale'
 
 export default function CreateBlockSalePage() {
