+++ web/src/components/block-sale/stand-form.tsx
