+++ web/src/components/block-sale/stand-panel.tsx
