+++ web/src/components/booking/tables-card.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, screen, fireEvent } from '@testing-library/react'
-import { BookableTablesCard } from './bookable-tables-card'
+import { BookableTablesCard } from './tables-card'
 import type { MarketTable } from '@fyndstigen/shared'
 
 vi.mock('@/lib/flags', () => ({ useFlag: () => true, getFlagEnv: () => true }))
