+++ web/src/components/market/image-gallery.tsx
