+++ web/src/components/route-builder/route-builder.tsx
@@ -3,12 +3,12 @@
 import Link from 'next/link'
 import { useRouteBuilder } from '@/hooks/use-route-builder'
 import { useAuth } from '@/lib/auth/auth-context'
-import { FyndstigenLogo } from './fyndstigen-logo'
-import { RouteFormFields } from './route-builder/route-form-fields'
-import { StopList, type RouteBuilderStop } from './route-builder/stop-list'
-import { RouteMap } from './route-builder/route-map'
-import { SaveRouteButton } from './route-builder/save-route-button'
-import { AnonSaveForm } from './route-builder/anon-save-form'
+import { FyndstigenLogo } from '../fyndstigen-logo'
+import { RouteFormFields } from './route-form-fields'
+import { StopList, type RouteBuilderStop } from './stop-list'
+import { RouteMap } from './route-map'
+import { SaveRouteButton } from './save-route-button'
+import { AnonSaveForm } from './anon-save-form'
 import type { FleaMarketNearBy } from '@fyndstigen/shared'
 import type { OpeningHourRule, OpeningHourException } from '@fyndstigen/shared'
